refactor(types): move VirtualNodeMap to vnode.rs
